fixing some warnings

+++ horde/g_horde.cpp
@@ -3442,7 +3442,7 @@ void Horde_CleanBodies()
 	}
 
 	// Clean up gibs, bodyque, and stuck projectiles
-	for (int i = game.maxclients + 1; i < globals.num_edicts; i++)
+	for (int i = game.maxclients + 1; i < static_cast<int>(globals.num_edicts); i++)
 	{
 		edict_t* ent = &g_edicts[i];
